data type correction

+++ lectures-palestras/recursividade.c
@@ -4,19 +4,19 @@
 #include <ctype.h>
 #include <string.h>
 
-double fatorial(int x); 
+long fatorial(int x); 
 
 int main(void)
 {
     int num = get_int("Digite um numero: ");
     
 
-    printf("O fatorial de %i e %d\n", num, fatorial(num));
+    printf("O fatorial de %i e %ld\n", num, fatorial(num));
 
     return 0;
 }
 
- double fatorial(int x)
+ long fatorial(int x)
     {
         if(x == 1 || x == 0)
         {
